Move QPSUT0x sample problems to qpsolvers.problems
